Change ll to int

+++ Segment Tree/segmentTree3.cpp
@@ -7,23 +7,23 @@ template <typename Type>
 struct segment_tree
 {
 private:
-	ll n1;
+	int n1;
 	vector<Type> tree;
 
 public:
-	void build(ll n, Type *a)
+	void build(int n, Type *a)
 	{
 		tree.resize(4 * n);
 		n1 = 1;
 		while (n1 < n)
 			n1 *= 2;
 		n1--;
-		for (ll i = 1; i <= n; i++)
+		for (int i = 1; i <= n; i++)
 			tree[i + n1] = a[i];
-		for (ll i = n1; i >= 1; i--)
+		for (int i = n1; i >= 1; i--)
 			tree[i] = tree[i * 2] + tree[i * 2 + 1];
 	}
-	void upd(ll pos, Type k)
+	void upd(int pos, Type k)
 	{
 		pos += n1;
 		tree[pos] += k;
@@ -31,7 +31,7 @@ public:
 		for (; pos; pos /= 2)
 			tree[pos] = tree[pos * 2] + tree[pos * 2 + 1];
 	}
-	Type sum(ll l, ll r)
+	Type sum(int l, int r)
 	{
 		l += n1;
 		r += n1;
